refactor: move layout components into separate folder and extract inner leyout method
