+++ components/InnerLayout/InnerLayout.tsx
@@ -1,60 +1,23 @@
 /** @jsxRuntime classic */
 /** @jsx jsx */
 import React from 'react'
-import { ThemeProvider, jsx } from 'theme-ui'
+import { ThemeProvider, Themed, jsx } from 'theme-ui'
 import dynamic from 'next/dynamic'
-import { ManagedUIContext, useUI } from '@components/ui/context'
-import { Head, Navbar } from '@components/common'
+import { useUI } from '@components/ui/context'
 import { useAcceptCookies } from '@lib/hooks/useAcceptCookies'
 import { Button } from 'theme-ui'
 import { Sidebar } from '@components/ui'
 import { CartSidebarView } from '@components/cart'
-import { CommerceProvider } from '@lib/shopify/storefront-data-hooks'
-import shopifyConfig from '@config/shopify'
-import { builder, BuilderContent, Builder } from '@builder.io/react'
+import { builder, Builder } from '@builder.io/react'
 import themesMap from '@config/theme'
 import '@builder.io/widgets'
 import 'react-spring-modal/styles.css'
-import seoConfig from '@config/seo.json'
-import NoSSR from './NoSSR'
+import NoSSR from '../common/NoSSR'
+import Header from '@components/Header/Header'
 
 const FeatureBar = dynamic(() => import('@components/common/FeatureBar'), {
   ssr: false,
 })
-
-const Layout: React.FC<{ pageProps: any }> = ({ children, pageProps }) => {
-  const builderTheme = pageProps.theme
-  const isLive = !Builder.isEditing && !Builder.isPreviewing
-  return (
-    <CommerceProvider {...shopifyConfig}>
-      <BuilderContent
-        isStatic
-        {...(isLive && { content: builderTheme })}
-        modelName="theme"
-      >
-        {(data, loading) => {
-          if (loading && !builderTheme) {
-            return 'loading ...'
-          }
-          const siteSettings = data?.siteSettings
-          const colorOverrides = data?.colorOverrides
-          const siteSeoInfo = data?.siteInformation
-          return (
-            <ManagedUIContext key={data?.id} siteSettings={siteSettings}>
-              <Head seoInfo={siteSeoInfo || seoConfig} />
-              <InnerLayout
-                themeName={data?.theme || 'base'}
-                colorOverrides={colorOverrides}
-              >
-                {children}
-              </InnerLayout>
-            </ManagedUIContext>
-          )
-        }}
-      </BuilderContent>
-    </CommerceProvider>
-  )
-}
 
 const InnerLayout: React.FC<{
   themeName: string
@@ -75,20 +38,22 @@ const InnerLayout: React.FC<{
   }
   const { displaySidebar, closeSidebar } = useUI()
   const { acceptedCookies, onAcceptCookies } = useAcceptCookies()
+  const { logo, headerNavigation } = useUI()
   return (
     <ThemeProvider theme={theme}>
-      <Navbar />
-      <div
+      <Header navigation={headerNavigation} logo={logo} theme={theme} />
+      <Themed.div
         sx={{
           margin: `0 auto`,
           px: 20,
           maxWidth: 1920,
           minWidth: '60vw',
           minHeight: 800,
+          mt: '92px',
         }}
       >
         <main>{children}</main>
-      </div>
+      </Themed.div>
 
       <Sidebar
         open={
@@ -113,4 +78,4 @@ const InnerLayout: React.FC<{
   )
 }
 
-export default Layout
+export default InnerLayout
