+++ components/common/index.ts
@@ -1,6 +1,4 @@
 export { default as FeatureBar } from './FeatureBar'
-export { default as Layout } from './Layout'
-export { default as Navbar } from './Navbar'
 export { default as Searchbar } from './Searchbar'
 export { default as UserNav } from './UserNav'
 export { default as Head } from './Head'
